requirements.txt
Django==5.0.2,
Pillow==10.2.0,
psycopg2-binary==2.9.9,
django-redis==5.4.0,
PyJWT==2.8.0,
djangorestframework==3.14.0,
fuzzywuzzy==0.18.0,
Levenshtein==0.25.0